requirements.txt
fastapi==0.111.1
uvicorn==0.30.6
python-multipart==0.0.17
pydantic==2.9.2
requests==2.32.3
beautifulsoup4==4.12.3
PyPDF2==3.0.1
python-dotenv==1.0.1
openai==1.53.0
qdrant-client==1.12.1
sentence-transformers==3.2.1
ell-ai==0.0.15
pytesseract==0.3.13
python-jose==3.3.0
tabula-py==2.10.0
PyMuPDF==1.24.13
python-dateutil==2.9.0.post0
jpype1==1.5.1
spacy==3.8.2
en_core_web_trf==3.8.0